High-Pass filter scrapped
Filters should be applied before any work is done on the audio file and the filtered data should be treated as the original data.
Therefore, it is better to have a separate Activity to handle filters and save the result as a new audio file.
Also, high-pass filters might not suffice as bass tones might also be more important than the melody in some cases.